+++ src/main/java/pl/emilweglowski/ui/gui/reservations/AddNewReservationScene.java
@@ -47,7 +47,7 @@ public class AddNewReservationScene {
         List<RoomSelectionItem> roomSelectionItems = new ArrayList<>();
 
         roomsAsDTO.forEach(dto -> {
-            roomSelectionItems.add(new RoomSelectionItem(dto.getNumber(), dto.getId()));
+            roomSelectionItems.add(new RoomSelectionItem(dto.getNumber(), (int)dto.getId()));
         });
 
         List<GuestSelectionItem> guestSelectionItems = new ArrayList<>();
